requirements.txt
python-telegram-bot[job-queue]==22.5
psycopg2-binary~=2.9.11
